requirements.txt
requests
python-firebase
pyautogui
pywin32
win32gui